backend/pyproject.toml
[tool.poetry]
name = "pileup-buster-backend"
version = "1.0.0"
description = "Ham radio callsign queue management system API"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pymongo = "^4.5.0"
python-dotenv = "^1.0.0"
dnspython = "^2.4.2"
httpx = "^0.25.2"
callsignlookuptools = "^1.1.1"
requests = "^2.32.4"
playwright = "^1.40.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

# add entry points for CLI commands
[tool.poetry.scripts]
start = "app.app:main"
